HT_19/sears/requirements/development.txt
-r base.txt
django-extensions==3.2.3
ipython==8.19.0
isort==5.13.2
flake8==6.1.0